fix bottom calculation for DrawText, clean up
